funcion para cambiar imagen de perfil

+++ src/components/UserReview/CardReview.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { Review } from "./ReviewList";
 import { useAuth } from "@/contexts/authContext";
 import { deleteReview } from "@/services/reviewService";
+import Swal from "sweetalert2";
 
 interface CardReviewProps {
     review: Review;
@@ -13,12 +14,27 @@ const CardReview: React.FC<CardReviewProps> = ({ review, onDelete }) => {
     const { dataUser } = useAuth();
 
     const handleDelete = async (id: string) => {
-        try {
-            await deleteReview(id);
-            console.log("Review deleted successfully");
-            onDelete(id); // Llama a la función para actualizar la lista en el componente padre
-        } catch (error) {
-            console.error("Error deleting review:", error);
+        // Mostrar el mensaje de confirmación con Swal
+        const result = await Swal.fire({
+            title: "Are you sure?",
+            text: "Do you really want to delete this review? This action cannot be undone.",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: "#3085d6",
+            confirmButtonText: "Yes, delete it!"
+        });
+
+        // Si el usuario confirma, eliminar la reseña
+        if (result.isConfirmed) {
+            try {
+                await deleteReview(id);
+                onDelete(id); // Llama a la función para actualizar la lista en el componente padre
+                Swal.fire("Deleted!", "Your review has been deleted.", "success");
+            } catch (error) {
+                console.error("Error deleting review:", error);
+                Swal.fire("Error", "There was a problem deleting the review.", "error");
+            }
         }
     };
 
@@ -26,8 +42,8 @@ const CardReview: React.FC<CardReviewProps> = ({ review, onDelete }) => {
         <div className="p-4 bg-gray-100 rounded-xl border-gray-300 border shadow-md flex flex-col space-y-4">
             <div className="flex justify-between items-center">
                 <img
-                    src={dataUser?.user.image}
-                    alt="Profile Picture"
+                    src={review.user.image}
+                    alt="Profile"
                     className="w-12 h-12 rounded-full object-cover"
                 />
                 <h3 className="text-lg font-semibold">{review.user.name}</h3>
